% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_qpcr_step_1.R
\name{calculate_qpcr_step_1}
\alias{calculate_qpcr_step_1}
\title{Calculates delta CT for each sample.}
\usage{
calculate_qpcr_step_1(df, target_now, housekeeping_now)
}
\arguments{
\item{df}{input tibble in wide format}

\item{target_now}{target gene}

\item{housekeeping_now}{reference gene}
}
\value{
A tibble with delta CT values
}
\description{
Calculates delta CT for each sample.
}
