% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_qpcr_step_2.R
\name{calculate_qpcr_step_2}
\alias{calculate_qpcr_step_2}
\title{Calculates delta-delta CT and fold-change for each sample.}
\usage{
calculate_qpcr_step_2(
  df,
  group_vars,
  filter_var,
  ctrl_condition,
  summarise_vars
)
}
\arguments{
\item{df}{input tibble in wide format with deltaCT}

\item{group_vars}{grouping variable(s), quoted}

\item{filter_var}{filtering variable, quoted}

\item{ctrl_condition}{control condition for filtering variable equality test}

\item{summarise_vars}{variable(s) to summarise, quoted}
}
\value{
A tibble with expression fold-changes
}
\description{
Calculates delta-delta CT and fold-change for each sample.
}
