% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mybarplot.R
\name{plot_mybarplot}
\alias{plot_mybarplot}
\title{Plot barplot with data points and error bars}
\usage{
plot_mybarplot(
  df,
  x_var = "concentration",
  y_var = "expression_FC",
  facet_var = "treatment",
  fill_by = "treatment",
  summary_stat = "mean_se",
  label = FALSE
)
}
\arguments{
\item{df}{input data set as tibble}

\item{x_var}{aesthetic for plotting}

\item{y_var}{aesthetic for plotting}

\item{facet_var}{variable for faceting, quoted}

\item{fill_by}{variable for fill, quoted}

\item{summary_stat}{summary statistic for geom_errorbar}

\item{label}{include text label to indicate summary_stat}
}
\value{
A barplot with data points and error bars
}
\description{
Plot barplot with data points and error bars
}
