% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qpcr_target.R
\name{plot_qpcr_target}
\alias{plot_qpcr_target}
\title{QC plot for qPCR data}
\usage{
plot_qpcr_target(df, facet_vars = c("target", "condition"), color_by, pch_by)
}
\arguments{
\item{df}{input data set as tibble, normalized to control condition}

\item{facet_vars}{variables for faceting, quoted}

\item{color_by}{variable for coloring, quoted}

\item{pch_by}{variable for plotting character, quoted}
}
\value{
A plot of 2^(CT_ctrl - CT) for each target and condition
}
\description{
QC plot for qPCR data
}
