% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_control.R
\name{summarise_control}
\alias{summarise_control}
\title{Title}
\usage{
summarise_control(df, group_vars, filter_var, ctrl_condition, summarise_vars)
}
\arguments{
\item{df}{input data set as tibble}

\item{group_vars}{grouping variable(s), quoted}

\item{filter_var}{filtering variable, quoted}

\item{ctrl_condition}{control condition for filtering variable equality test}

\item{summarise_vars}{variable(s) to summarise, quoted}
}
\value{
A tibble of mean control values and sd
}
\description{
Title
}
